+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 #  Changelog
 
+- 4.3.3
+    - Add support for Swift Package Manager (SPM)
+    - Make `reuseIdentifier` public accessible
+    - Fix warnings
+- 4.3.2
+    - Fixed a bug where editing immediately before a tag prefix would break the tag
+- 4.3.1
+    - Add `RxSwift`/`RxCocoa` support through extensions and delegate proxies, requires Cocoapods installation of  `InputBarAccessoryView/RxExtensions`
 - 4.3.0
     - Swift 5 support
     - Fixes accessibility of HorizontalEdgePadding initializers and a typo in its filename